- changed test package name from org.baeldung to com.baeldung - streams are added where neccessary - format fixes
