+++ spring-apache-camel/pom.xml
@@ -11,7 +11,7 @@
 	<properties>
 		<env.camel.version>2.16.1</env.camel.version>
 		<env.spring.version>4.3.4.RELEASE</env.spring.version>
-        <maven-surefire-plugin.version>2.19.1</maven-surefire-plugin.version>
+		<maven-surefire-plugin.version>2.19.1</maven-surefire-plugin.version>
 		<java.version>1.8</java.version>
 		<junit.version>4.12</junit.version>
 	</properties>
@@ -48,7 +48,11 @@
 			<artifactId>spring-context</artifactId>
 			<version>${env.spring.version}</version>
 		</dependency>
-
+		<dependency>
+			<groupId>org.slf4j</groupId>
+			<artifactId>slf4j-log4j12</artifactId>
+			<version>1.7.21</version>
+		</dependency>
 	</dependencies>
 
 	<build>
