+++ spring-apache-camel/src/main/resources/camel-context-MulticastFileRouterTest.xml
@@ -4,7 +4,7 @@
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.1.xsd
 	 http://camel.apache.org/schema/spring http://camel.apache.org/schema/spring/camel-spring.xsd">
 
-	<bean id="multicastFileRouter" class="org.baeldung.camel.file.MulticastFileRouter" />
+	<bean id="multicastFileRouter" class="com.baeldung.camel.file.MulticastFileRouter" />
 
 	<camelContext xmlns="http://camel.apache.org/schema/spring">
 		<routeBuilder ref="multicastFileRouter" />
