+++ spring-apache-camel/src/main/resources/camel-context-SplitterFileRouter.xml
@@ -4,7 +4,7 @@
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.1.xsd
 	 http://camel.apache.org/schema/spring http://camel.apache.org/schema/spring/camel-spring.xsd">
 
-	<bean id="splitterFileRouter" class="org.baeldung.camel.file.SplitterFileRouter" />
+	<bean id="splitterFileRouter" class="com.baeldung.camel.file.SplitterFileRouter" />
 
 	<camelContext xmlns="http://camel.apache.org/schema/spring">
 		<routeBuilder ref="splitterFileRouter" />
