+++ spring-apache-camel/src/test/java/com/apache/camel/file/processor/ContentBasedFileRouterIntegrationTest.java
@@ -1,32 +1,35 @@
-package org.apache.camel.file.processor;
+package com.apache.camel.file.processor;
 
 import java.io.File;
 
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
-public class MulticastFileRouterIntegrationTest {
+@RunWith(JUnit4.class)
+public class ContentBasedFileRouterIntegrationTest {
 
     private static final long DURATION_MILIS = 10000;
     private static final String SOURCE_FOLDER = "src/test/source-folder";
-    private static final String DESTINATION_FOLDER_WORLD = "src/test/destination-folder-world";
-    private static final String DESTINATION_FOLDER_HELLO = "src/test/destination-folder-hello";
+    private static final String DESTINATION_FOLDER_TXT = "src/test/destination-folder-txt";
+    private static final String DESTINATION_FOLDER_OTHER = "src/test/destination-folder-other";
 
     @Before
     public void setUp() throws Exception {
         File sourceFolder = new File(SOURCE_FOLDER);
-        File destinationFolderWorld = new File(DESTINATION_FOLDER_WORLD);
-        File destinationFolderHello = new File(DESTINATION_FOLDER_HELLO);
+        File destinationFolderTxt = new File(DESTINATION_FOLDER_TXT);
+        File destinationFolderOther = new File(DESTINATION_FOLDER_OTHER);
 
         cleanFolder(sourceFolder);
-        cleanFolder(destinationFolderWorld);
-        cleanFolder(destinationFolderHello);
+        cleanFolder(destinationFolderTxt);
+        cleanFolder(destinationFolderOther);
 
         sourceFolder.mkdirs();
         File file1 = new File(SOURCE_FOLDER + "/File1.txt");
-        File file2 = new File(SOURCE_FOLDER + "/File2.txt");
+        File file2 = new File(SOURCE_FOLDER + "/File2.csv");
         file1.createNewFile();
         file2.createNewFile();
     }
@@ -45,7 +48,7 @@ public class MulticastFileRouterIntegrationTest {
     @Test
     @Ignore
     public void routeTest() throws InterruptedException {
-        ClassPathXmlApplicationContext applicationContext = new ClassPathXmlApplicationContext("camel-context-MulticastFileRouterTest.xml");
+        ClassPathXmlApplicationContext applicationContext = new ClassPathXmlApplicationContext("camel-context-ContentBasedFileRouterTest.xml");
         Thread.sleep(DURATION_MILIS);
         applicationContext.close();
