+++ spring-apache-camel/src/test/java/com/apache/camel/main/AppIntegrationTest.java
@@ -1,4 +1,4 @@
-package org.apache.camel.main;
+package com.apache.camel.main;
 
 import com.baeldung.camel.main.App;
 import junit.framework.TestCase;
